UI Fix Again
Yeah basiclly school is a throw away day beacuse of rain and snow so why
not code? And by code I mean do the one coding homework I had to do that
did not involve coding. I fixed the king becuase his picture started
freaking out and I changed the pieces to fit on the tiles.