Rooks! Also a "minor" grafical glitch.
Rooks were added to the game, however a big, not game breaking but
annoying glitch was discovered. When the New Game Button is hit, all of
the pieces fo back to their original places, but the other half of the
board that is ment to be empty does not become empty again. Sometimes it
will clear after a turn, and sometimes it will not clear at all until
the program is restarted.