Glitch Fix!
First, don't worry about the time this was published.
Second, I fixed the glitch.
The glitch happened because after the king realized it was in check, it
set w/bKingX and w/bKingY to x1 and y1, respectivly. Then when you
clicked the submit button again, it would check if w/bKingX and w/bKingY
were equal to x1 and y1, and since they were just set to be equal t
those, it went down the if statement and then set w/bKingX and w/bKingY
to x2 and y2, respecitvly. Since the player was moving a different
piece, the board registered the king at a different tile, and would not
register check.
The best part about this was that debug did not work, which I learned
after clicking next line through two 8 long for loops inside another 8
for loop, which took a lot of clicking, so I jsut kinda had to stare at
the code for a solid 15 min until I relazied the problem.

Overall, I rate Java a solid 10/10.