Pawn Promotion Implemented
-If a pawn is moved to the 1st or 8th rank, then the user must input a
capital letter in the destination box at the 4th character position to
indicate the desired promotion piece. Example: "h8=Q". This is parsed to
a Piece.TypePiece and is passed to board.move(), which passes it to
board.executeMove(). If the desired promotion piece is not indicated,
the move is rejected as invalid.