A mistake
The way our check detector was designed, we ignored all other pieces and
assumed there were no pieces between the attacker and the king. While
this works fine for the pawn and the knight, it doesn't work so well
when the game notifies me that I am in check becuase my queen and the
enemy king share the same row. It'll be "fun" to code that in. Yay. I
also fixed a couple typos with me putting the wrong variables, got rid
of the substantial amount of lag, and fixed a sign change that was
causing the lag.