An overestimate
I was wrong. After realizing I made even more sign mistakes, this time
with more than and less than, I thought of a solution. It wasn't very
hard, just had to check in the while loop if the space it was currently
searching was occupied. Since the while loop will immediately quit out
if it scannign the enemy king's square, it will not detect the king as
an obstruction.