Additions for captured pieces but also a fat bug
The rest of the pieces now have captured images, however I noticed a big
error. Whenever you move a piece to the last rank, (for each side so
a-h8 for white and a-h0 for black), it makes that piece into a pawn. If
you promote a pawn it shows up,(except for the fact i commented it out),
and then turns into a pawn next turn.