Found part of the pawn problem, added some error messages, and corrected more things in opposing check.
I found the issue with pieces turning into pawns on the last rank, its
on line 246 in ChessUI when promote is set to pawn. At least thats part
of the problem. I also added in a new variable, valid, which would be
set true if the move was valid, or set false if it was not. It is used
for error messages and tells you if the move is illegal or not. Also
found an error or two in opposing check which were fixed.