Added forfeit, draw, and move display.
I added the forfeit button and the draw button's code to forfeit and
draw. They don't do this yes because I am not sure how to forfeit or
draw the game without the game ending. I also added a little line of
code so the computer output's whose turn it is.